Disable Menu editing on customer facing page (#89)
What was done:
- Add checks to `MenuInteractor` to ensure that `isEditable` is true before mutating anything
- Remove drag support from collection view if `isEditable` is false.
- Menu should pass `isEditable` to MenuDetail.

As the category drop handler still handles drops, it's possible to drop menu items onto the categories on the customer page, e.g. by dragging and holding from the stall admin page and navigating back to the customer page before dropping it. However, this will not do anything as the interactor checks if `isEditable` is true.